# apps/test_app/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: Simple Todo REST API for the testing lifecycle demonstration
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server

tags:
  - name: todos
    description: Todo item operations
  - name: system
    description: System health

paths:
  /health:
    get:
      tags: [system]
      summary: Health check
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: todo-api

  /todos:
    get:
      tags: [todos]
      summary: List all todos
      operationId: listTodos
      parameters:
        - name: completed
          in: query
          required: false
          schema:
            type: boolean
          description: Filter by completion status
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'

    post:
      tags: [todos]
      summary: Create a new todo
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreate'
      responses:
        '201':
          description: Todo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  /todos/{todo_id}:
    get:
      tags: [todos]
      summary: Get a todo by ID
      operationId: getTodo
      parameters:
        - $ref: '#/components/parameters/TodoId'
      responses:
        '200':
          description: The todo item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'

    put:
      tags: [todos]
      summary: Update a todo
      operationId: updateTodo
      parameters:
        - $ref: '#/components/parameters/TodoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
      responses:
        '200':
          description: Updated todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    delete:
      tags: [todos]
      summary: Delete a todo
      operationId: deleteTodo
      parameters:
        - $ref: '#/components/parameters/TodoId'
      responses:
        '204':
          description: Todo deleted
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'

components:
  parameters:
    TodoId:
      name: todo_id
      in: path
      required: true
      schema:
        type: string
      description: The unique identifier of the todo item

  schemas:
    TodoCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Buy groceries
        description:
          type: string
          maxLength: 1000
          nullable: true
          example: Milk, bread, eggs
        completed:
          type: boolean
          default: false

    TodoUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
          nullable: true
        completed:
          type: boolean

    Todo:
      type: object
      required: [id, title, completed, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          example: Buy groceries
        description:
          type: string
          nullable: true
          example: Milk, bread, eggs
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
          example: Todo 'abc' not found

    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string